fix: BufferedSMBReader.read returns Int32, not Int
The compiler is strict: read(_:size:) -> Int32, but the window-serving
path returned the raw Int lastServed. Cast it.

+++ ios/Runner/BufferedSMBReader.swift
@@ -90,7 +90,7 @@ final class BufferedSMBReader: IOReader, @unchecked Sendable {
             _ = serveIfResident(into: buffer, max: Int(size))
         }
         cond.unlock()
-        return lastServed
+        return Int32(lastServed)
     }
 
     private var lastServed = 0
